backend/requirements.txt
alembic==1.14.1
email-validator==2.2.0
fastapi>=0.135.0
greenlet==3.2.4
httpx==0.28.1
pydantic-settings==2.7.0
psycopg[binary]==3.2.13
python-dotenv==1.0.1
SQLAlchemy==2.0.45
pypdf>=4.0.0
uvicorn[standard]>=0.32.1
